refactor: Dashboard.vue to AppHome.vue
